chore(wip): remove circular dependency by introducing common-type-file (relates #1626 #1627)
