+++ src/lib/core/capability.test.ts
@@ -7,11 +7,11 @@ import { expect, describe, jest, beforeEach, it } from "@jest/globals";
 import { Event, Operation } from "../enums";
 import { PeprMutateRequest } from "../mutate-request";
 import { PeprValidateRequest } from "../validate-request";
-import { AdmissionRequest } from "../types";
 import { WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
 import { GenericClass } from "kubernetes-fluent-client";
 import { Schedule } from "./schedule";
 import { OnSchedule } from "./schedule";
+import { AdmissionRequest } from "../common-types";
 
 // Mocking isBuildMode, isWatchMode, and isDevMode globally
 jest.mock("./envChecks", () => ({
