+++ src/lib/filter/filter.test.ts
@@ -27,7 +27,7 @@ import {
   filterNoMatchReason,
   shouldSkipRequest,
 } from "./filter";
-import { AdmissionRequest, Binding } from "../types";
+import { Binding } from "../types";
 import { Event, Operation } from "../enums";
 import {
   defaultAdmissionRequest,
@@ -35,6 +35,7 @@ import {
   defaultFilters,
   defaultKubernetesObject,
 } from "./adjudicators/defaultTestObjects";
+import { AdmissionRequest } from "../common-types";
 
 const callback = () => undefined;
