+++ src/lib/mutate-request.test.ts
@@ -3,9 +3,9 @@
 
 import { beforeEach, describe, expect, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { AdmissionRequest } from "./types";
 import { Operation } from "./enums";
 import { PeprMutateRequest } from "./mutate-request";
+import { AdmissionRequest } from "./common-types";
 
 describe("PeprMutateRequest", () => {
   let mockRequest: AdmissionRequest<KubernetesObject>;
