+++ src/lib/mutate-request.ts
@@ -1,10 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { AdmissionRequest, DeepPartial } from "./types";
 import { Operation } from "./enums";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { clone, mergeDeepRight } from "ramda";
+import { AdmissionRequest } from "./common-types";
+import { DeepPartial } from "./utility-types";
 
 // PeprMutateRequest class for mutation request handling
 export class PeprMutateRequest<T extends KubernetesObject> {
