+++ src/lib/processors/mutate-processor.ts
@@ -8,7 +8,7 @@ import { MeasureWebhookTimeout } from "../telemetry/webhookTimeouts";
 import { Capability } from "../core/capability";
 import { shouldSkipRequest } from "../filter/filter";
 import { MutateResponse } from "../k8s";
-import { AdmissionRequest, Binding } from "../types";
+import { Binding } from "../types";
 import Log from "../telemetry/logger";
 import { ModuleConfig } from "../types";
 import { PeprMutateRequest } from "../mutate-request";
@@ -17,6 +17,7 @@ import { OnError } from "../../cli/init/enums";
 import { resolveIgnoreNamespaces } from "../assets/webhooks";
 import { Operation } from "fast-json-patch";
 import { WebhookType } from "../enums";
+import { AdmissionRequest } from "../common-types";
 export interface Bindable {
   req: AdmissionRequest;
   config: ModuleConfig;
