+++ src/lib/validate-request.test.ts
@@ -3,9 +3,9 @@
 
 import { beforeEach, describe, expect, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { ValidateActionResponse, AdmissionRequest } from "./types";
 import { Operation } from "./enums";
 import { PeprValidateRequest } from "./validate-request";
+import { AdmissionRequest, ValidateActionResponse } from "./common-types";
 describe("PeprValidateRequest", () => {
   let mockRequest: AdmissionRequest<KubernetesObject>;
